V1.5.8
Added: Menu before starting Pong game, with an additional back to menu button at the top of the screen. Also added a Game over screen which does not have styling at the moment but will display when the game reaches 10 games. Also added a ParticleJS Background to fit with the rest of the theme.