V1.6.4
Changed: Indexing for Visibility and UV index which now actually give you real information, altered the way the data is displayed to show proper formats such as % and added a date calculator to allow me to call any daily data easier in the future. Styling wise adjusted the columns to be able to hold this much data and work in mobile and also gave the wind direction additional styling to mimic a compass with its direction.